# cmd/docs/swagger.yaml
definitions:
  github_com_kangyueyue_short-url_internal_http.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.PShortUrlData:
    properties:
      clientId:
        description: 用户id
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      expireAt:
        description: 过期时间
        type: string
      id:
        type: integer
      longUrl:
        description: 长链-使用utf8mb4_bin字符集，区分大小写
        type: string
      shortUrl:
        description: 短链-使用utf8mb4_bin字符集，区分大小写
        type: string
      updatedAt:
        type: string
    type: object
  vo.DelVo:
    properties:
      isDisk:
        description: 是否删除磁盘文件
        type: boolean
      shortUrl:
        type: string
    required:
    - shortUrl
    type: object
  vo.GetVO:
    properties:
      shortUrl:
        type: string
    required:
    - shortUrl
    type: object
  vo.SetVo:
    properties:
      clientId:
        type: string
      longUrls:
        items:
          type: string
        type: array
    required:
    - clientId
    - longUrls
    type: object
info:
  contact: {}
paths:
  /hello:
    get:
      consumes:
      - application/json
      description: 这是一个Hello World接口
      produces:
      - application/json
      responses:
        "200":
          description: Hello World
          schema:
            type: string
      summary: Hello 示例接口
      tags:
      - 测试
  /short_url/del:
    delete:
      consumes:
      - application/json
      description: 根据短链标识和客户端ID删除对应的短链
      parameters:
      - description: 客户端ID
        in: header
        name: client_id
        required: true
        type: string
      - description: 删除请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vo.DelVo'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_kangyueyue_short-url_internal_http.Response'
      summary: 删除短链
      tags:
      - 短链接
  /short_url/get:
    post:
      consumes:
      - application/json
      description: 根据短链接获取原始长链接
      parameters:
      - description: 请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vo.GetVO'
      - description: 客户端ID
        in: header
        name: client_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回长链接信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kangyueyue_short-url_internal_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PShortUrlData'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_kangyueyue_short-url_internal_http.Response'
      summary: 获取长链接
      tags:
      - 短链接
  /short_url/list:
    get:
      consumes:
      - application/json
      description: 根据client_id获取对应的短链列表
      parameters:
      - description: 客户端ID
        in: header
        name: client_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 短链列表
          schema:
            items:
              $ref: '#/definitions/models.PShortUrlData'
            type: array
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_kangyueyue_short-url_internal_http.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_kangyueyue_short-url_internal_http.Response'
      summary: 获取短链列表
      tags:
      - 短链接
  /short_url/set:
    post:
      consumes:
      - application/json
      description: 批量创建短链接，将长链接转换为短链接
      parameters:
      - description: 请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vo.SetVo'
      - description: 客户端ID
        in: header
        name: client_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回短链接信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kangyueyue_short-url_internal_http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PShortUrlData'
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_kangyueyue_short-url_internal_http.Response'
      summary: 创建短链接
      tags:
      - 短链接
swagger: "2.0"